+++ sdk/src/test/java/com/hedera/hashgraph/sdk/FreezeTransactionTest.java
@@ -27,9 +27,8 @@ import org.junit.jupiter.api.Test;
 import org.threeten.bp.Instant;
 
 import java.util.Arrays;
-import java.util.Collections;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class FreezeTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
@@ -71,6 +70,6 @@ public class FreezeTransactionTest {
     void shouldBytes() throws Exception {
         var tx = spawnTestTransaction();
         var tx2 = FreezeTransaction.fromBytes(tx.toBytes());
-        assertEquals(tx.toString(), tx2.toString());
+        assertThat(tx2.toString()).isEqualTo(tx.toString());
     }
 }