+++ sdk/src/test/java/com/hedera/hashgraph/sdk/TokenBurnTransactionTest.java
@@ -28,7 +28,7 @@ import org.threeten.bp.Instant;
 import java.util.Arrays;
 import java.util.Collections;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class TokenBurnTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
@@ -86,13 +86,13 @@ public class TokenBurnTransactionTest {
     void shouldBytesFungible() throws Exception {
         var tx = spawnTestTransaction();
         var tx2 = TokenBurnTransaction.fromBytes(tx.toBytes());
-        assertEquals(tx.toString(), tx2.toString());
+        assertThat(tx2.toString()).isEqualTo(tx.toString());
     }
 
     @Test
     void shouldBytesNft() throws Exception {
         var tx = spawnTestTransactionNft();
         var tx2 = TokenBurnTransaction.fromBytes(tx.toBytes());
-        assertEquals(tx.toString(), tx2.toString());
+        assertThat(tx2.toString()).isEqualTo(tx.toString());
     }
 }